style buttons inline flex-start
